@sudden6 2019-01-14
@anthonybilinski I didn't actually setup perf for this, but I'm quite sure it's now faster than before. Maybe you can redo your tests with these patches.