@anthonybilinski 2019-01-15
Unfortunately my slow laptop is won't start its display manager at the moment after an update, but testing on my desktop and with perf:
before change, consistently 10s from "login" to window
after change, consistently 5s from "login" to window
before change, emojis disabled, consistently 5s  from "login" to window

I'm missing symbols in my libs, but here's the perf run from before change, with emojis. The emoji parsing is the wide stack starting at _ZN14QMe.
![image](https://user-images.githubusercontent.com/10469203/51191719-dab70080-1899-11e9-9b56-5410b791b1d8.png)

After change, with emojis enabled, where you can see that call stack is much thinner:
![image](https://user-images.githubusercontent.com/10469203/51191750-ef939400-1899-11e9-81d4-6d3b4ee80647.png)

Since on my ARM laptop before the problem went away completely disabling smileys, I'm guessing I'll see 3m40 -> 5s instead of just a halving of time on my desktop, I guess our other start-up times are less CPU bound.

Nice fix!