@sudden6 2019-01-15
Thanks for the tests, with that reduction I'm now pretty confident it will do for your ARM Chromebook too. Else we can still implement the other things I mentioned in the Issue.

> why is this a slot rather than a private member function?

was a mistake, fixed it